datasource with annotation
