+++ src/main/java/com/kids/crm/todo/repository/TodoRepository.java
@@ -1,36 +1,13 @@
 package com.kids.crm.todo.repository;
 
 import com.kids.crm.todo.model.Todo;
-import com.kids.crm.todo.utils.PersistenceManager;
 
-import javax.persistence.EntityManager;
-import javax.persistence.Query;
+import javax.ejb.Local;
+import javax.ejb.TransactionAttribute;
 import java.util.List;
 
-public class TodoRepository {
-    public void createTodo(Todo todo){
-        EntityManager em = PersistenceManager.INSTANCE.getEntityManager();
-        em.getTransaction()
-                .begin();
-        em.persist(todo);
-        em.getTransaction()
-                .commit();
-        em.close();
-        PersistenceManager.INSTANCE.close();
-    }
-
-    public List<Todo> fetch() {
-        EntityManager em = PersistenceManager.INSTANCE.getEntityManager();
-        em.getTransaction()
-                .begin();
-
-        Query query = em.createQuery("select t from Todo t");
-        List<Todo> todos = query.getResultList();
-        em.getTransaction()
-                .commit();
-        em.close();
-        PersistenceManager.INSTANCE.close();
-        return todos;
-
-    }
+@TransactionAttribute
+@Local
+public interface TodoRepository {
+    List<Todo> fetch();
 }
