+++ src/main/resources/META-INF/persistence.xml
@@ -2,7 +2,7 @@
              xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/persistence
 http://xmlns.jcp.org/xml/ns/persistence/persistence_2_1.xsd"
              version="2.1">
-    <persistence-unit name="jpa-example" transaction-type="RESOURCE_LOCAL">
+    <persistence-unit name="jpa-example" transaction-type="JTA">
         <provider>org.hibernate.jpa.HibernatePersistenceProvider</provider>
         <properties>
             <property name="javax.persistence.jdbc.url" value="jdbc:mysql://localhost/todo" />
